+++ ui-lib/src/main/java/co/electriccoin/zcash/di/ProviderModule.kt
@@ -3,7 +3,10 @@ package co.electriccoin.zcash.di
 import co.electriccoin.zcash.ui.common.provider.GetDefaultServersProvider
 import co.electriccoin.zcash.ui.common.provider.GetVersionInfoProvider
 import co.electriccoin.zcash.ui.common.provider.GetZcashCurrencyProvider
+import co.electriccoin.zcash.ui.common.provider.LocalAddressBookStorageProvider
+import co.electriccoin.zcash.ui.common.provider.LocalAddressBookStorageProviderImpl
 import org.koin.core.module.dsl.factoryOf
+import org.koin.dsl.bind
 import org.koin.dsl.module
 
 val providerModule =
@@ -11,4 +14,5 @@ val providerModule =
         factoryOf(::GetDefaultServersProvider)
         factoryOf(::GetVersionInfoProvider)
         factoryOf(::GetZcashCurrencyProvider)
+        factoryOf(::LocalAddressBookStorageProviderImpl) bind LocalAddressBookStorageProvider::class
     }