+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/repository/AddressBookRepository.kt
@@ -1,12 +1,21 @@
 package co.electriccoin.zcash.ui.common.repository
 
+import co.electriccoin.zcash.ui.common.datasource.LocalAddressBookDataSource
+import co.electriccoin.zcash.ui.common.datasource.RemoteAddressBookProvider
+import co.electriccoin.zcash.ui.common.model.AddressBook
 import co.electriccoin.zcash.ui.common.model.AddressBookContact
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.NonCancellable
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.onSubscription
 import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
+import kotlinx.coroutines.withContext
 
 interface AddressBookRepository {
-    val contacts: StateFlow<List<AddressBookContact>>
+    val addressBook: Flow<AddressBook?>
 
     suspend fun saveContact(
         name: String,
@@ -20,18 +29,31 @@ interface AddressBookRepository {
     )
 
     suspend fun deleteContact(contact: AddressBookContact)
-
-    suspend fun getContact(id: String): AddressBookContact?
 }
 
-class AddressBookRepositoryImpl : AddressBookRepository {
-    override val contacts = MutableStateFlow(emptyList<AddressBookContact>())
+class AddressBookRepositoryImpl(
+    private val localAddressBookDataSource: LocalAddressBookDataSource,
+    private val remoteAddressBookProvider: RemoteAddressBookProvider
+) : AddressBookRepository {
+    private val semaphore = Mutex()
+    private val addressBookCache = MutableStateFlow<AddressBook?>(null)
+
+    override val addressBook: Flow<AddressBook?> =
+        addressBookCache
+            .onSubscription {
+                withNonCancellableSemaphore {
+                    ensureSynchronization()
+                }
+            }
 
     override suspend fun saveContact(
         name: String,
         address: String
-    ) {
-        contacts.update { it + AddressBookContact(name = name.trim(), address = address.trim()) }
+    ) = withNonCancellableSemaphore {
+        ensureSynchronization()
+        val local = localAddressBookDataSource.saveContact(name, address)
+        addressBookCache.update { local }
+        remoteAddressBookProvider.uploadContacts()
     }
 
     override suspend fun updateContact(
@@ -39,27 +61,46 @@ class AddressBookRepositoryImpl : AddressBookRepository {
         name: String,
         address: String
     ) {
-        contacts.update {
-            it.toMutableList()
-                .apply {
-                    set(
-                        it.indexOf(contact),
-                        AddressBookContact(name = name.trim(), address = address.trim())
-                    )
-                }
-                .toList()
+        withNonCancellableSemaphore {
+            ensureSynchronization()
+            val local = localAddressBookDataSource.updateContact(contact, name, address)
+            addressBookCache.update { local }
+            remoteAddressBookProvider.uploadContacts()
         }
     }
 
-    override suspend fun deleteContact(contact: AddressBookContact) {
-        contacts.update {
-            contacts.value.toMutableList()
-                .apply {
-                    remove(contact)
-                }
-                .toList()
+    override suspend fun deleteContact(contact: AddressBookContact) =
+        withNonCancellableSemaphore {
+            ensureSynchronization()
+            val local = localAddressBookDataSource.deleteContact(contact)
+            addressBookCache.update { local }
+            remoteAddressBookProvider.uploadContacts()
+        }
+
+    private suspend fun ensureSynchronization() {
+        if (addressBookCache.value == null) {
+            val merged =
+                mergeContacts(
+                    local = localAddressBookDataSource.getContacts(),
+                    remote = remoteAddressBookProvider.fetchContacts(),
+                )
+
+            localAddressBookDataSource.saveContacts(merged)
+            remoteAddressBookProvider.uploadContacts()
+
+            addressBookCache.update { merged }
         }
     }
 
-    override suspend fun getContact(id: String): AddressBookContact? = contacts.value.find { it.id == id }
+    @Suppress("UNUSED_PARAMETER")
+    private fun mergeContacts(
+        local: AddressBook,
+        remote: AddressBook?
+    ): AddressBook = local // TBD
+
+    private suspend fun withNonCancellableSemaphore(block: suspend () -> Unit) {
+        withContext(NonCancellable + Dispatchers.Default) {
+            semaphore.withLock { block() }
+        }
+    }
 }