+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/usecase/ObserveAddressBookContactsUseCase.kt
@@ -1,9 +1,10 @@
 package co.electriccoin.zcash.ui.common.usecase
 
 import co.electriccoin.zcash.ui.common.repository.AddressBookRepository
+import kotlinx.coroutines.flow.map
 
 class ObserveAddressBookContactsUseCase(
     private val addressBookRepository: AddressBookRepository
 ) {
-    operator fun invoke() = addressBookRepository.contacts
+    operator fun invoke() = addressBookRepository.addressBook.map { it?.contacts }
 }