+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/usecase/ValidateContactAddressUseCase.kt
@@ -3,6 +3,8 @@ package co.electriccoin.zcash.ui.common.usecase
 import co.electriccoin.zcash.ui.common.model.AddressBookContact
 import co.electriccoin.zcash.ui.common.repository.AddressBookRepository
 import co.electriccoin.zcash.ui.common.repository.WalletRepository
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.first
 
 class ValidateContactAddressUseCase(
     private val addressBookRepository: AddressBookRepository,
@@ -15,7 +17,7 @@ class ValidateContactAddressUseCase(
         val result = walletRepository.getSynchronizer().validateAddress(address)
         return when {
             result.isNotValid -> Result.Invalid
-            addressBookRepository.contacts.value
+            addressBookRepository.addressBook.filterNotNull().first().contacts
                 .filter {
                     if (exclude == null) true else it != exclude
                 }