+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 ## [Unreleased]
 
 ### Changed
+- Scan QR screen has been redesigned
 - The Receive screen UI has been redesigned
 - Send screen redesigned & added a possibility to pick a contact from address book
 - Confirmation screen redesigned & added a contact name to the transaction if the contact is in address book